Move sans IO panda client to crypto submodule
